+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 ## main
-- Added a exception when the file is not found in a estate file query
+
+## v0.7.3
+- Added an exception when the file is not found in a estate file query
+- Added a modify method for estate files
 
 ## v0.7.2
 - Fixed a typo where the delete method of estate files returns an error